assignment_deadline_env/requirements.txt
openenv-core
pydantic>=2.0.0
fastapi
uvicorn
python-dotenv
openai
